Refactor Home Agent BE service
